+++ src/Utils/Constants.py
@@ -1,0 +1,8 @@
+SUMMARY_COLUMNS = ['Datetime (UTC)','Timezone (minutes)','Unix Timestamp (UTC)','Acc magnitude avg','Eda avg','Temp avg','Movement intensity','Steps count','Rest','On Wrist']
+TIME_SERIES_COLUMNS = ['Acc magnitude avg','Eda avg','Temp avg','Movement intensity','Steps count','Rest']
+
+SUMMARYFILENAME = 'summary.csv'
+METADATAFILENAME = 'metadata.csv'
+APPNAME='Digi-Vis'
+APPVERSION = '0.0.0.0'
+DEFAULTPATHTODATASETS = './datasets'
