+++ src/Window.py
@@ -2,17 +2,17 @@ import tkinter as tk
 from tkinter.ttk import Notebook
 from src.Frame.DataBuilderFrame import DataBuilderFrame
 
-APPNAME='Digi-Vis'
-
 class Window(tk.Tk):
-    def __init__(self, pathToDatasets):
+    def __init__(self, appname, pathToDatasets):
         super().__init__()
         self.geometry('800x600')
-        self.title(APPNAME)
+        self.title(appname)
 
         self.notebook = Notebook(self)
 
-        data_builder_tab = DataBuilderFrame(pathToDatasets)
+        data_builder_tab = DataBuilderFrame(self, pathToDatasets)
         self.notebook.add(data_builder_tab, text='Builder')
         self.notebook.pack(fill=tk.BOTH, expand=1)
 
+    def AddTab(self, tab, text):
+        self.notebook.add(tab, text=text)
